PYIC-1718: Rename the validate-cri-check lambda
