+++ deploy/template.yaml
@@ -1280,19 +1280,19 @@ Resources:
       FilterPattern: ""
       LogGroupName: !Ref IPVProcessJourneyStepFunctionLogGroup
 
-  IPVValidateCriCheckFunction:
+  ValidateCriCredentialFunction:
     Type: AWS::Serverless::Function
     DependsOn:
-      - "IPVValidateCriCheckFunctionLogGroup"
+      - "ValidateCriCredentialFunctionLogGroup"
     Properties:
       # checkov:skip=CKV_AWS_115: We do not have enough data to allocate the concurrent execution allowance per function.
       # checkov:skip=CKV_AWS_116: Lambdas invoked via API Gateway do not support Dead Letter Queues.
       # checkov:skip=CKV_AWS_117: Lambdas will migrate to our own VPC in future work.
-      FunctionName: !Sub "validate-cri-check-${Environment}"
-      Handler: uk.gov.di.ipv.core.validatecricheck.ValidateCriCheckHandler::handleRequest
+      FunctionName: !Sub "validate-cri-credential-${Environment}"
+      Handler: uk.gov.di.ipv.core.validatecricredential.ValidateCriCredentialHandler::handleRequest
       Runtime: java11
       PackageType: Zip
-      CodeUri: ../lambdas/validatecricheck
+      CodeUri: ../lambdas/validate-cri-credential
       Architectures:
         - arm64
       MemorySize: 2048
@@ -1301,7 +1301,7 @@ Resources:
         # checkov:skip=CKV_AWS_173: These environment variables do not require encryption.
         Variables:
           ENVIRONMENT: !Ref Environment
-          POWERTOOLS_SERVICE_NAME: !Sub validate-cri-check-${Environment}
+          POWERTOOLS_SERVICE_NAME: !Sub validate-cri-credential-${Environment}
           IPV_SESSIONS_TABLE_NAME: !Ref SessionsTable
           USER_ISSUED_CREDENTIALS_TABLE_NAME: !Ref UserIssuedCredentialsV2Table
       Policies:
@@ -1326,20 +1326,20 @@ Resources:
           - ProvisionedConcurrentExecutions: !FindInMap [ EnvironmentConfiguration, !Ref AWS::AccountId, provisionedConcurrency ]
           - !Ref AWS::NoValue
 
-  IPVValidateCriCheckFunctionLogGroup:
+  ValidateCriCredentialFunctionLogGroup:
     Type: AWS::Logs::LogGroup
     Properties:
       RetentionInDays: 14
-      LogGroupName: !Sub "/aws/lambda/validate-cri-check-${Environment}"
+      LogGroupName: !Sub "/aws/lambda/validate-cri-credential-${Environment}"
       KmsKeyId: !GetAtt LoggingKmsKey.Arn
 
-  IPVValidateCriCheckFunctionLogGroupSubscriptionFilter:
+  ValidateCriCredentialFunctionLogGroupSubscriptionFilter:
     Type: AWS::Logs::SubscriptionFilter
     Condition: IsNotDevelopment
     Properties:
       DestinationArn: "arn:aws:logs:eu-west-2:885513274347:destination:csls_cw_logs_destination_prod"
       FilterPattern: ""
-      LogGroupName: !Ref IPVValidateCriCheckFunctionLogGroup
+      LogGroupName: !Ref ValidateCriCredentialFunctionLogGroup
 
   EvaluateGpg45ScoresFunction:
     Type: AWS::Serverless::Function
