+++ lambdas/validate-cri-credential/build.gradle
