+++ lambdas/validate-cri-credential/src/main/java/uk/gov/di/ipv/core/validatecricredential/domain/KbvEvidence.java
@@ -1,0 +1,17 @@
+package uk.gov.di.ipv.core.validatecricredential.domain;
+
+import uk.gov.di.ipv.core.library.annotations.ExcludeFromGeneratedCoverageReport;
+
+@ExcludeFromGeneratedCoverageReport
+public class KbvEvidence {
+
+    private final int verificationScore;
+
+    public KbvEvidence(int verificationScore) {
+        this.verificationScore = verificationScore;
+    }
+
+    public int getVerificationScore() {
+        return verificationScore;
+    }
+}
