+++ lambdas/validate-cri-credential/src/main/java/uk/gov/di/ipv/core/validatecricredential/validation/CriEvidenceValidator.java
@@ -1,0 +1,7 @@
+package uk.gov.di.ipv.core.validatecricredential.validation;
+
+import com.nimbusds.jose.shaded.json.JSONArray;
+
+public interface CriEvidenceValidator {
+    boolean validate(JSONArray evidenceArray);
+}
