+++ lambdas/validate-cri-credential/src/main/resources/IpvLambdaJsonLayout.json
