+++ lambdas/validate-cri-credential/src/main/resources/log4j2.yaml
