+++ lambdas/validate-cri-credential/src/test/java/uk/gov/di/ipv/core/validatecricredential/validators/KbvEvidenceValidatorTest.java
@@ -1,0 +1,35 @@
+package uk.gov.di.ipv.core.validatecricredential.validators;
+
+import org.junit.jupiter.api.Test;
+import uk.gov.di.ipv.core.library.exceptions.HttpResponseExceptionWithErrorBody;
+import uk.gov.di.ipv.core.library.persistence.item.UserIssuedCredentialsItem;
+import uk.gov.di.ipv.core.validatecricredential.validation.CriCheckValidator;
+
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static uk.gov.di.ipv.core.library.fixtures.TestFixtures.SIGNED_KBV_VC_FAILED;
+import static uk.gov.di.ipv.core.library.fixtures.TestFixtures.SIGNED_KBV_VC_PASSED;
+import static uk.gov.di.ipv.core.validatecricredential.validation.CriCheckValidator.CRI_ID_KBV;
+
+class KbvEvidenceValidatorTest {
+    private final CriCheckValidator criCheckValidator = new CriCheckValidator();
+
+    @Test
+    void returnsTrueForSuccessfulKbvCheck() throws HttpResponseExceptionWithErrorBody {
+        UserIssuedCredentialsItem userIssuedCredentialsItem = new UserIssuedCredentialsItem();
+        userIssuedCredentialsItem.setCredential(SIGNED_KBV_VC_PASSED);
+        userIssuedCredentialsItem.setCredentialIssuer(CRI_ID_KBV);
+
+        assertTrue(criCheckValidator.isSuccess(userIssuedCredentialsItem));
+    }
+
+    @Test
+    void returnsFalseForKbvCheckWithLowVerificationScore()
+            throws HttpResponseExceptionWithErrorBody {
+        UserIssuedCredentialsItem userIssuedCredentialsItem = new UserIssuedCredentialsItem();
+        userIssuedCredentialsItem.setCredential(SIGNED_KBV_VC_FAILED);
+        userIssuedCredentialsItem.setCredentialIssuer(CRI_ID_KBV);
+
+        assertFalse(criCheckValidator.isSuccess(userIssuedCredentialsItem));
+    }
+}
