--- lambdas/validatecricheck/src/main/java/uk/gov/di/ipv/core/validatecricheck/domain/KbvEvidence.java
@@ -1,17 +1,0 @@
-package uk.gov.di.ipv.core.validatecricheck.domain;
-
-import uk.gov.di.ipv.core.library.annotations.ExcludeFromGeneratedCoverageReport;
-
-@ExcludeFromGeneratedCoverageReport
-public class KbvEvidence {
-
-    private final int verificationScore;
-
-    public KbvEvidence(int verificationScore) {
-        this.verificationScore = verificationScore;
-    }
-
-    public int getVerificationScore() {
-        return verificationScore;
-    }
-}
