--- lambdas/validatecricheck/src/main/java/uk/gov/di/ipv/core/validatecricheck/validation/CriEvidenceValidator.java
@@ -1,7 +1,0 @@
-package uk.gov.di.ipv.core.validatecricheck.validation;
-
-import com.nimbusds.jose.shaded.json.JSONArray;
-
-public interface CriEvidenceValidator {
-    boolean validate(JSONArray evidenceArray);
-}
